chore(swaps): Separate BoltzSwapProvider into partials
